pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name        = "christman-crypto"
version     = "1.0.0"
description = "Seven-tier hybrid cryptographic stack — Vigenère through post-quantum ML-KEM. The Christman AI Project."
readme      = "README.md"
license     = { file = "LICENSE" }
authors     = [{ name = "Everett Christman", email = "everett@christmanaai.org" }]
keywords    = [
    "cryptography", "post-quantum", "ml-kem", "kyber", "fips-203",
    "aes-256-gcm", "chacha20", "xchacha20", "rsa", "steganography",
    "hybrid-encryption", "christman"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies    = [
    "cryptography>=41.0",
]

[project.optional-dependencies]
steg = ["Pillow>=10.0"]
kyber = ["kyber-py>=0.3"]
all  = ["Pillow>=10.0", "kyber-py>=0.3"]
dev  = ["pytest>=7.0", "Pillow>=10.0"]

[project.urls]
Homepage   = "https://github.com/EverettNC/Harvest-Now-Decrypt-Later"
Repository = "https://github.com/EverettNC/Harvest-Now-Decrypt-Later"

[tool.setuptools.packages.find]
where = ["."]
include = ["christman_crypto*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
